requirements.txt
numpy==1.24.3
pandas==1.5.3
tqdm==4.66.5
matplotlib==3.7.1
mplsoccer==1.2.2
plotly==5.9.0
gspread
google-auth
numpy
requests
beautifulsoup4